Making macro calls more robust.
Macros now properly respect parenthesis when parsing arguments, so
CALL(a,(b,c),d) will parse as three arguments:
   1. a
   2. b,c
   3. d

Additionally, if a macro is called with more arguments then it
expects, any arguments in excess of the expected amount are
concatenated onto the final argument.

This may be a bit unusual, but it allows embedding complicated
macro expressions without worrying about proper parenthetical
grouping if the macro only has one argument.

For example, in LaTeX it is common to have things like:
LATEX_BLOCK(
  \begin{align*}
  \vec{V}_t &= <o_x,o_y,o_z> + <d_x, d_y, d_z> t
  \end{align*}
)

Which would otherwise have to be grouped in extra parenthesis
to accomodate the commas in the vector literals.